requirements.txt
python-telegram-bot
together
pillow
python-dotenv